Fix compilation warnings

+++ src/frame.h
@@ -249,19 +249,19 @@ public:
     operator bool() const { return isValid() && isComplete(); }
 
     uint8_t *data(size_t plane = 0) {
-        if (!m_raw || plane >= (AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
+        if (!m_raw || plane >= size_t(AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
             return nullptr;
         return m_raw->extended_data[plane];
     }
 
     const uint8_t *data(size_t plane = 0) const {
-        if (!m_raw || plane >= (AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
+        if (!m_raw || plane >= size_t(AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
             return nullptr;
         return m_raw->extended_data[plane];;
     }
 
     size_t size(size_t plane) const {
-        if (!m_raw || plane >= (AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
+        if (!m_raw || plane >= size_t(AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
             return 0;
         AVBufferRef *buf = plane < AV_NUM_DATA_POINTERS ?
                                m_raw->buf[plane] :
